Fix #4: Split endpoints
