+++ project/__init__.py
@@ -31,7 +31,9 @@ def create_app():
     db.init_app(app)
 
     # register blueprints
-    from project.api.views import dx_blueprint
-    app.register_blueprint(dx_blueprint)
+    from project.api.views.items import items_blueprint
+    from project.api.views.items_statuses import items_statuses_blueprint
+    app.register_blueprint(items_blueprint)
+    app.register_blueprint(items_statuses_blueprint)
 
     return app
