+++ project/api/views/items.py
@@ -1,17 +1,14 @@
-# project/api/views.py
-
-
 from flask import Blueprint, jsonify, request
 
-from project.api.models import Item, Item_Status
+from project.api.models import Item
 from project import db
 
 from sqlalchemy import exc
 
-dx_blueprint = Blueprint('items', __name__)
+items_blueprint = Blueprint('items', __name__)
 
 
-@dx_blueprint.route('/ping', methods=['GET'])
+@items_blueprint.route('/ping', methods=['GET'])
 def ping_pong():
     return jsonify({
         'status': 'success',
@@ -19,7 +16,7 @@ def ping_pong():
     })
 
 
-@dx_blueprint.route('/items', methods=['POST'])
+@items_blueprint.route('/items', methods=['POST'])
 def add_item():
     post_data = request.get_json()
     if not post_data:
@@ -91,7 +88,7 @@ def add_item():
         return jsonify(response_object), 400
 
 
-@dx_blueprint.route('/items/<item_id>', methods=['GET'])
+@items_blueprint.route('/items/<item_id>', methods=['GET'])
 def get_single_item(item_id):
     """Get single item details"""
     response_object = {
@@ -122,7 +119,7 @@ def get_single_item(item_id):
         return jsonify(response_object), 404
 
 
-@dx_blueprint.route('/items', methods=['GET'])
+@items_blueprint.route('/items', methods=['GET'])
 def get_all_items():
     """Get all items"""
     items = Item.query.all()
@@ -149,7 +146,7 @@ def get_all_items():
     return jsonify(response_object), 200
 
 
-@dx_blueprint.route('/items/<item_id>', methods=['PATCH'])
+@items_blueprint.route('/items/<item_id>', methods=['PATCH'])
 def edit_single_item():
     """Edit a single todo item"""
     post_data = request.get_json()
@@ -234,130 +231,3 @@ def edit_single_item():
             'message': 'Invalid payload.'
         }
         return jsonify(response_object), 400
-
-
-@dx_blueprint.route('/item_statuses', methods=['POST'])
-def add_item_status():
-    post_data = request.get_json()
-    if not post_data:
-        response_object = {
-            'status': 'fail',
-            'message': 'Invalid payload.'
-        }
-        return jsonify(response_object), 400
-    name = post_data.get('name')
-    value = post_data.get('value')
-    value_type = post_data.get('value_type')
-
-    if not name:
-        response_object = {
-            'status': 'fail',
-            'message': 'Invalid payload.'
-        }
-        return jsonify(response_object), 400
-
-    if not value:
-        response_object = {
-            'status': 'fail',
-            'message': 'Invalid payload.'
-        }
-        return jsonify(response_object), 400
-
-    if not value_type:
-        response_object = {
-            'status': 'fail',
-            'message': 'Invalid payload.'
-        }
-        return jsonify(response_object), 400
-
-
-    try:
-        Item_Status.query.filter_by(name=name).first()
-        item_status = Item_Status.query.filter_by(name=name).first()
-
-        if not item_status:
-            item_status = db.session.add(Item_Status(name=name, value=value, value_type=value_type))
-            db.session.commit()
-            item_status = Item_Status.query.filter_by(name=name).first()
-            response_object = {
-                'status': 'success',
-                'message': '{} was added!'.format(name),
-                'data': {
-                    'id': item_status.id,
-                    'name': item_status.name,
-                    'value': item_status.value,
-                    'value_type': item_status.value_type,
-                    'created_at': item_status.created_at,
-                    'updated_at': item_status.updated_at
-                }
-            }
-            return jsonify(response_object), 201
-        else:
-            response_object = {
-                'status': 'fail',
-                'message': 'Sorry. That item status already exists.'
-            }
-            return jsonify(response_object), 400
-    except exc.IntegrityError as e:
-        db.session.rollback()
-        response_object = {
-            'status': 'fail',
-            'message': 'Invalid payload.'
-        }
-        return jsonify(response_object), 400
-
-
-@dx_blueprint.route('/item_statuses/<item_status_id>', methods=['GET'])
-def get_single_item_status(item_status_id):
-    """Get single item status details"""
-    response_object = {
-        'status': 'fail',
-        'message': 'Item status does not exist'
-    }
-    try:
-        item_status = Item_Status.query.filter_by(id=int(item_status_id)).first()
-        if not item_status:
-            return jsonify(response_object), 404
-        else:
-            response_object = {
-                'status': 'success',
-                'data': {
-                    'id' : item_status_id,
-                    'name': item_status.name,
-                    'value': item_status.value,
-                    'value_type': item_status.value_type,
-                    'created_at': item_status.created_at,
-                    'updated_at': item_status.updated_at
-                }
-            }
-            return jsonify(response_object), 200
-    except ValueError:
-        return jsonify(response_object), 404
-
-
-@dx_blueprint.route('/item_statuses', methods=['GET'])
-def get_all_item_statuses():
-    """Get all item statuses"""
-    items_statuses = Item_Status.query.all()
-    item_status_list = []
-    for item_status in items_statuses:
-        item_status_object = {
-            'id': item_status.id,
-            'name': item_status.name,
-            'value': item_status.value,
-            'value_type': item_status.value_type,
-            'created_at': item_status.created_at,
-            'updated_at': item_status.updated_at
-        }
-        item_status_list.append(item_status_object)
-    response_object = {
-        'status': 'success',
-        'data': {
-            'item_statuses': item_status_list
-        }
-    }
-    return jsonify(response_object), 200
-
-@dx_blueprint.route('/item_statuses/<item_status_id>', methods=['PATCH'])
-def edit_single_item_status():
-    """Edit a single todo item"""
